+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -131,6 +131,7 @@ class PlatformInAppWebViewWidgetCreationParams
       super.onContentSizeChanged,
       super.onProcessFailed,
       super.onAcceleratorKeyPressed,
+      super.onShowFileChooser,
       super.initialUrlRequest,
       super.initialFile,
       super.initialData,